More work on local sync

+++ src/main/java/org/mozilla/android/sync/repositories/bookmarks/LocalBookmarkSynchronizer.java
@@ -84,7 +84,7 @@ public class LocalBookmarkSynchronizer {
     while (curMoz.isAfterLast() == false) {
       
       // Find bookmark in android store
-      long androidId = curMoz.getLong(curMoz.getColumnIndex(BookmarksDatabaseHelper.COL_ANDROID_ID));
+      int androidId = curMoz.getInt(curMoz.getColumnIndex(BookmarksDatabaseHelper.COL_ANDROID_ID));
       String where = Browser.BookmarkColumns._ID + "=" + androidId;
       Cursor curDroid = context.getContentResolver().query(Browser.BOOKMARKS_URI, null, where, null, null);
       curDroid.moveToFirst();
@@ -146,8 +146,6 @@ public class LocalBookmarkSynchronizer {
     
     return true;
   }
-  
-
   
   // Create new moz bookmark from droid cursor
   // containing title, url, id
